+++ src/main/java/dk/kb/datahandler/api/v1/impl/DsDatahandlerApiServiceImpl.java
@@ -165,5 +165,15 @@ public class DsDatahandlerApiServiceImpl extends ImplBase implements DsDatahandl
         }
     }
 
+    @Override
+    public void kalturaDeltaUpload(Long mTimeFrom) {    
+        try {
+           DsDatahandlerFacade.kalturaDeltaUpload(mTimeFrom);
+        }
+        catch(Exception e) {
+            throw handleException(e);
+        }                       
+    }
+
 
 }